#33: Document new classes and split off Layer interface
